requirements.txt
cffi==2.0.0
cryptography==46.0.3
dotenv==0.9.9
mysql-connector-python==9.5.0
pycparser==2.23
python-dotenv==1.2.1
